rename calendar to ZetkinCalendar
